Seperating Physical discovery with Physical system descriptor
