+++ tests/ttnn/unit_tests/gtests/multiprocess/test_send_recv_pipeline.cpp
@@ -14,6 +14,7 @@
 #include <tt-metalium/experimental/sockets/mesh_socket.hpp>
 #include <tt-metalium/distributed_context.hpp>
 #include "tt_metal/fabric/physical_system_descriptor.hpp"
+#include "tt_metal/fabric/physical_system_discovery.hpp"
 
 namespace tt::tt_metal {
 
@@ -140,10 +141,8 @@ PhysicalSystemDescriptor create_physical_system_descriptor() {
     const auto& distributed_context = tt::tt_metal::MetalContext::instance().get_distributed_context_ptr();
     const auto& hal = tt::tt_metal::MetalContext::instance().hal();
     const auto& rtoptions = tt::tt_metal::MetalContext::instance().rtoptions();
-    constexpr bool run_discovery = true;
-    const auto& driver = cluster.get_driver();
-
-    return tt::tt_metal::PhysicalSystemDescriptor(driver, distributed_context, &hal, rtoptions, run_discovery);
+    auto& driver_ref = const_cast<tt::umd::Cluster&>(*cluster.get_driver());
+    return run_physical_system_discovery(driver_ref, distributed_context, &hal, rtoptions.get_target_device());
 }
 
 // This test does the following:
