+++ tools/scaleout/validation/run_cluster_validation.cpp
@@ -13,6 +13,7 @@
 #include <cxxopts.hpp>
 #include <factory_system_descriptor/utils.hpp>
 #include "tt_metal/fabric/physical_system_descriptor.hpp"
+#include "tt_metal/fabric/physical_system_discovery.hpp"
 #include <tt-metalium/distributed.hpp>
 #include "tt_metal/impl/context/metal_context.hpp"
 #include <cabling_generator/cabling_generator.hpp>
@@ -282,11 +283,11 @@ PhysicalSystemDescriptor generate_physical_system_descriptor(const InputArgs& in
         return physical_system_descriptor;
     }
     log_output_rank0("Running Physical Discovery");
-    constexpr bool run_discovery = true;
     auto& context = tt::tt_metal::MetalContext::instance();
     const auto& driver = context.get_cluster().get_driver();
-    auto physical_system_descriptor = tt::tt_metal::PhysicalSystemDescriptor(
-        driver, context.get_distributed_context_ptr(), &context.hal(), context.rtoptions(), run_discovery);
+    auto& driver_ref = const_cast<tt::umd::Cluster&>(*driver);
+    auto physical_system_descriptor = run_physical_system_discovery(
+        driver_ref, context.get_distributed_context_ptr(), &context.hal(), context.rtoptions().get_target_device());
     log_output_rank0("Physical Discovery Complete");
     log_output_rank0("Detected Hosts: " + log_hostnames(physical_system_descriptor.get_all_hostnames()));
     return physical_system_descriptor;
@@ -385,7 +386,18 @@ int main(int argc, char* argv[]) {
         reset_ethernet_links(physical_system_descriptor, missing_asic_topology);
         links_reset = true;
         num_retrains++;
-        physical_system_descriptor.run_discovery(true, true);
+        // Re-run discovery
+        auto& context_ref = tt::tt_metal::MetalContext::instance();
+        physical_system_descriptor.clear();
+        auto& driver_ref = const_cast<tt::umd::Cluster&>(*context_ref.get_cluster().get_driver());
+        auto new_psd = run_physical_system_discovery(
+            driver_ref,
+            context_ref.get_distributed_context_ptr(),
+            &context_ref.hal(),
+            context_ref.rtoptions().get_target_device(),
+            true,
+            true);
+        physical_system_descriptor.merge(std::move(new_psd));
         missing_asic_topology = run_connectivity_validation(input_args, physical_system_descriptor);
     }
