+++ tt_metal/fabric/serialization/physical_system_descriptor_serialization.cpp
@@ -235,12 +235,7 @@ std::unique_ptr<PhysicalSystemDescriptor> proto_to_physical_system_descriptor(
     if (!target_device_type.has_value()) {
         throw std::runtime_error("Invalid target device type: " + std::to_string(proto_desc.target_device_type()));
     }
-    auto descriptor = std::make_unique<PhysicalSystemDescriptor>(
-        PhysicalSystemDescriptor::null_cluster,
-        nullptr,
-        nullptr,
-        *target_device_type,
-        false);  // Don't run discovery
+    auto descriptor = std::make_unique<PhysicalSystemDescriptor>(*target_device_type);
 
     // Convert system graph
     auto& system_graph = descriptor->get_system_graph();
