+++ tt_metal/fabric/sources.cmake
@@ -60,6 +60,7 @@ set(FABRIC_SOURCES
     serialization/intermesh_connections_serialization.cpp
     serialization/port_descriptor_serialization.cpp
     ccl/ccl_common.cpp
+    physical_system_discovery.cpp
     physical_system_descriptor.cpp
     topology_mapper.cpp
     topology_mapper_utils.cpp
